A Modification to Add a $titlecase script function.
This:
"the word is EAT THIS NOW. USA. USB. US VS us. the big bad fire truck is going to eat you! abba & ac/dc are playiNG on a Farm down by the river"

Becomes:
The Word Is Eat This Now. Usa. Usb. US vs Us. The Big Bad Fire Truck Is Going to Eat You! ABBA & AC/DC Are Playing on a Farm Down by the River.